Refactor Backend/Frontend structure & New AI model
