+++ backend/app/core/database.py
@@ -6,7 +6,7 @@ Road Report Backend - Database Module
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, AsyncSession
 from sqlalchemy.orm import DeclarativeBase
 
-from app.config import settings
+from app.core.config import settings
 
 
 # สร้าง Async Engine สำหรับ SQLite
