+++ frontend/src/services/api.js
@@ -1,0 +1,2 @@
+export const BASE_URL = "http://127.0.0.1:8000";
+export const API_REPORTS = `${BASE_URL}/api/reports`;
